requirements.txt
requests
playwright
PyYAML